requirements.txt
Django==3.2.2
Pillow==8.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
django-heroku==0.3.1